manager.php & global debugging

+++ api/public/app/Request.php
@@ -17,7 +17,7 @@ class Request
     static public function set(Session $session): void
     {
         $query = "INSERT INTO `requests` SET `session` = ?, `executor` = ?, `code` = ?, `lang` = ?
-                       ON DUPLICATE KEY UPDATE `executor` = ?, `code` = ?, `lang` = ?";
+                       ON DUPLICATE KEY UPDATE `executor` = ?, `code` = ?, `lang` = ?, `received` = 0";
         Db::get()->exec($query, [
             $session->id, $session->executor, $session->code, $session->lang,
             $session->executor, $session->code, $session->lang,
@@ -38,13 +38,12 @@ class Request
         }
         $query = "SELECT `session`, `code`, `lang`, md5(`code`) as `hash` FROM `requests` WHERE `executor` = ?";
         $params = [$executor];
-        if ($lang !== null) {
-            $query = " AND `lang` = ?";
+        if ($lang !== null && $hash !== null) {
+            $query .= " AND `lang` = ? AND md5(`code`) = ?";
             $params[] = $lang;
-        }
-        if ($hash !== null) {
-            $query = " AND md5(`code`) = ?";
             $params[] = $hash;
+        } else {
+            $query .= " AND `received` = 0";
         }
         $query .= " LIMIT ?";
         $params[] = self::MAX_REQUESTS_FOR_EXECUTOR_PER_REQUEST;
@@ -56,8 +55,20 @@ class Request
         return $requests;
     }
 
+    static public function markReceived(string $executor, string $lang, string $hash): void
+    {
+        if (!Utils::isUuid($executor)) {
+            return;
+        }
+        $query = "UPDATE `requests` SET `received` = 1 WHERE `executor` = ? and `lang` = ? and md5(`code`) = ?";
+        Db::get()->exec($query, [$executor, $lang, $hash]);
+    }
+
     static public function remove(string $executor, string $lang, string $hash): void
     {
+        if (!Utils::isUuid($executor)) {
+            return;
+        }
         $query = "DELETE FROM `requests` WHERE `executor` = ? and `lang` = ? and md5(`code`) = ?";
         Db::get()->exec($query, [$executor, $lang, $hash]);
     }
